requirements.txt
streamlit
Pillow
opencv-python
numpy
reportlab
piexif
matplotlib